chore: StarRatingUISlider를 파일과 그룹으로 분리
