+++ BoxOffice/StarRating/StarRatingView.swift
@@ -126,18 +126,3 @@ class StarRatingView: UIView {
         }
     }
 }
-
-class StarRatingUISlider: UISlider {
-    override func beginTracking(_ touch: UITouch, with event: UIEvent?) -> Bool {
-        let width = self.frame.size.width
-        let tapPoint = touch.location(in: self)
-        let tapPercent = tapPoint.x / width
-        let newValue = self.maximumValue * Float(tapPercent)
-        
-        if newValue != self.value {
-            self.value = newValue
-        }
-        
-        return true
-    }
-}
